scanner/requirements.txt
python-magic
pefile
yara-python
ollama
